fix(#166): restructure creator-domain and creator-service to standard package layout
Address Oracle review feedback — both packages now use proper subdirectory
layout matching creator-provider:

- creator-domain: models/ moved into creator_domain/models/
  - Enables 'from creator_domain.models.X import Y' imports
  - pyproject.toml updated to find creator_domain package
  - test_stage.py updated to use creator_domain.models imports

- creator-service: all 16 modules moved into creator_service/
  - Enables 'from creator_service.X import Y' imports
  - pyproject.toml switched from py-modules to packages.find
  - All service modules updated: removed sys.path hacks, use
    creator_domain.models.* and relative imports
  - All 9 test files updated: proper creator_service.* imports

Test results (per-package): domain 8/8, service 154/155 (1 pre-existing),
provider 12/12.
