--- packages/creator-domain/__init__.py
@@ -1,1 +1,0 @@
-"""creator-domain package scaffold."""
