+++ packages/creator-domain/creator_domain/__init__.py
@@ -1,0 +1,1 @@
+"""creator-domain: shared domain models for short-form pipeline."""
