+++ packages/creator-domain/creator_domain/models/audio_artifact.py
