+++ packages/creator-domain/creator_domain/models/script_draft.py
