+++ packages/creator-domain/creator_domain/models/video_artifact.py
