+++ packages/creator-domain/creator_domain/models/visual_asset.py
