+++ packages/creator-domain/pyproject.toml
@@ -10,4 +10,4 @@ requires = ["setuptools>=68.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["models*"]
+include = ["creator_domain", "creator_domain.*"]
