+++ packages/creator-service/creator_service/stage_review_service.py
@@ -2,18 +2,10 @@
 
 from __future__ import annotations
 
-import importlib
-import sys
 from datetime import datetime, timezone
-from pathlib import Path
 from typing import Any, Protocol
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "creator-domain"))
-domain_models = importlib.import_module("models")
-RunStage = domain_models.RunStage
-REVIEW_STAGES = domain_models.REVIEW_STAGES
-can_transition = domain_models.can_transition
-
+from creator_domain.models import RunStage, REVIEW_STAGES, can_transition
 
 class StageReviewStorageBackend(Protocol):
     async def create_review(self, row: dict[str, Any]) -> dict[str, Any]:
