+++ packages/creator-service/creator_service/subtitle_service.py
@@ -1,7 +1,7 @@
-"""Audio service for artifact storage and retrieval.
+"""Subtitle service for artifact storage and retrieval.
 
-Manages audio artifacts generated for runs. Each run can have multiple audio
-artifacts. Audio artifacts are NOT scene-versioned; the latest artifact
+Manages subtitle artifacts generated for runs. Each run can have multiple subtitle
+artifacts. Subtitle artifacts are NOT scene-versioned; the latest artifact
 per run is the current one.
 
 Follows the same Protocol → InMemory → Service pattern as
@@ -10,16 +10,10 @@ VisualAssetService and ScriptService.
 
 from __future__ import annotations
 
-import sys
 from datetime import datetime, timezone
-from pathlib import Path
 from typing import Any, Protocol
 
-_DOMAIN_DIR = str(Path(__file__).resolve().parent.parent / "creator-domain")
-if _DOMAIN_DIR not in sys.path:
-    sys.path.insert(0, _DOMAIN_DIR)
-
-from models.audio_artifact import AudioArtifact  # type: ignore[reportMissingImports]  # noqa: E402, I001
+from creator_domain.models.subtitle_artifact import SubtitleArtifact
 
 
 # ---------------------------------------------------------------------------
@@ -27,9 +21,9 @@ from models.audio_artifact import AudioArtifact  # type: ignore[reportMissingImp
 # ---------------------------------------------------------------------------
 
 
-class AudioStorageBackend(Protocol):
+class SubtitleStorageBackend(Protocol):
     async def save_artifact(self, row: dict[str, Any]) -> dict[str, Any]:
-        """Persist an audio artifact row and return stored row with id assigned.
+        """Persist a subtitle artifact row and return stored row with id assigned.
 
         The storage backend allocates the next id and inserts the row.
         Callers MUST NOT include ``id`` in *row*; the returned dict
@@ -55,8 +49,8 @@ class AudioStorageBackend(Protocol):
 # ---------------------------------------------------------------------------
 
 
-class InMemoryAudioStorage:
-    """In-memory storage for audio artifacts."""
+class InMemorySubtitleStorage:
+    """In-memory storage for subtitle artifacts."""
 
     def __init__(self) -> None:
         self._artifacts: list[dict[str, Any]] = []
@@ -96,9 +90,9 @@ class InMemoryAudioStorage:
 # ---------------------------------------------------------------------------
 
 
-class AudioService:
-    def __init__(self, storage: AudioStorageBackend | None = None) -> None:
-        self.storage = storage if storage is not None else InMemoryAudioStorage()
+class SubtitleService:
+    def __init__(self, storage: SubtitleStorageBackend | None = None) -> None:
+        self.storage = storage if storage is not None else InMemorySubtitleStorage()
 
     # -- Create -------------------------------------------------------------
 
@@ -107,22 +101,22 @@ class AudioService:
         run_id: int,
         path: str,
         *,
+        format: str = "srt",
         model_used: str | None = None,
         provider_type: str | None = None,
-        voice: str | None = None,
-    ) -> AudioArtifact:
-        """Create a new audio artifact for a run.
+    ) -> SubtitleArtifact:
+        """Create a new subtitle artifact for a run.
 
-        Audio artifacts are NOT versioned. The latest artifact per run
+        Subtitle artifacts are NOT versioned. The latest artifact per run
         is the current one.
         """
         metadata = {}
+        if format is not None:
+            metadata["format"] = format
         if model_used is not None:
             metadata["model_used"] = model_used
         if provider_type is not None:
             metadata["provider_type"] = provider_type
-        if voice is not None:
-            metadata["voice"] = voice
 
         row = {
             "run_id": run_id,
@@ -135,19 +129,19 @@ class AudioService:
 
     # -- Reads --------------------------------------------------------------
 
-    async def get_latest(self, run_id: int) -> AudioArtifact | None:
-        """Get the most recent audio artifact for a run."""
+    async def get_latest(self, run_id: int) -> SubtitleArtifact | None:
+        """Get the most recent subtitle artifact for a run."""
         row = await self.storage.get_latest_by_run(run_id)
         if row is None:
             return None
         return self._row_to_artifact(row)
 
-    async def list_by_run(self, run_id: int) -> list[AudioArtifact]:
-        """List all audio artifacts for a run, newest first."""
+    async def list_by_run(self, run_id: int) -> list[SubtitleArtifact]:
+        """List all subtitle artifacts for a run, newest first."""
         rows = await self.storage.list_by_run(run_id)
         return [self._row_to_artifact(r) for r in rows]
 
-    async def get_by_id(self, artifact_id: int) -> AudioArtifact | None:
+    async def get_by_id(self, artifact_id: int) -> SubtitleArtifact | None:
         """Get a single artifact by id."""
         row = await self.storage.get_artifact(artifact_id)
         if row is None:
@@ -157,9 +151,9 @@ class AudioService:
     # -- Internal -----------------------------------------------------------
 
     @staticmethod
-    def _row_to_artifact(row: dict[str, Any]) -> AudioArtifact:
-        """Convert a storage row to an AudioArtifact domain model."""
-        return AudioArtifact.from_row(row)
+    def _row_to_artifact(row: dict[str, Any]) -> SubtitleArtifact:
+        """Convert a storage row to a SubtitleArtifact domain model."""
+        return SubtitleArtifact.from_row(row)
 
 
-audio_service = AudioService()
+subtitle_service = SubtitleService()
