+++ packages/creator-service/creator_service/visual_asset_service.py
@@ -11,16 +11,10 @@ VisualPlanService and ScriptService.
 from __future__ import annotations
 
 import asyncio
-import sys
 from datetime import datetime, timezone
-from pathlib import Path
 from typing import Any, Protocol
 
-_DOMAIN_DIR = str(Path(__file__).resolve().parent.parent / "creator-domain")
-if _DOMAIN_DIR not in sys.path:
-    sys.path.insert(0, _DOMAIN_DIR)
-
-from models.visual_asset import VisualAsset  # type: ignore[reportMissingImports]  # noqa: E402, I001
+from creator_domain.models.visual_asset import VisualAsset
 
 
 # ---------------------------------------------------------------------------
